requirements.txt
allure-pytest==2.9.43
pytest==6.2.4
pytest-ordering==0.6
pytest-rerunfailures==10.0
pytest-xdist==2.3.0
selenium==4.1.0